requirements.txt
fastapi
uvicorn
huggingface_hub
python-multipart
pypdf
langchain
openai
faiss-cpu
tiktoken
python-dotenv
sentence-transformers
